--- src/components/icon/warning.tsx
@@ -1,15 +1,0 @@
-export default function WarningIcon() {
-  return (
-    <svg
-      width="1em"
-      height="1em"
-      viewBox="0 0 100 100"
-      xmlns="http://www.w3.org/2000/svg"
-      style={{ display: 'inline', verticalAlign: 'middle' }}
-    >
-      <circle cx="50" cy="50" r="45" stroke="currentColor" strokeWidth="8" fill="none" />
-      <line x1="50" y1="20" x2="50" y2="60" stroke="currentColor" strokeWidth="8" />
-      <circle cx="50" cy="75" r="5" fill="currentColor" />
-    </svg>
-  );
-}
